requirements.txt
apache-airflow-providers-apache-kafka
apache-airflow-providers-apache-spark
apache-airflow-providers-slack
kafka-python
psycopg2-binary